Expose cli from lib
