+++ crates/wasvy_cli/src/lib.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "cli")]
+pub mod cli;
 pub mod command;
 pub mod dependency;
 pub mod editor;
